+++ src/components/App.js
@@ -1,5 +1,13 @@
+import { Outlet } from "react-router-dom";
+import Nav from "./Nav";
+
 function App() {
-  return(<h1>hi</h1>)
+  return (
+    <>
+      <Nav />
+      <Outlet />
+    </>
+  );
 }
 
 export default App;
