+++ src/components/Button.js
@@ -1,4 +1,4 @@
-import Link from "react-router-dom";
+import { Link } from "react-router-dom";
 import style from "./Button.module.css";
 import classNames from "classnames/bind";
