+++ src/pages/BoardPage.js
@@ -1,0 +1,9 @@
+const BoardPage = () => {
+  return (
+    <>
+      <div>BoardPage</div>
+    </>
+  );
+};
+
+export default BoardPage;
