+++ src/pages/HomePage.js
@@ -1,0 +1,5 @@
+const HomePage = () => {
+  return <div>HomePage</div>;
+};
+
+export default HomePage;
