+++ src/pages/LoginPage.js
@@ -1,0 +1,5 @@
+const LoginPage = () => {
+  return <div>LoginPage</div>;
+};
+
+export default LoginPage;
