+++ src/pages/NotFoundPage.js
@@ -1,0 +1,5 @@
+const NotFoundPage = () => {
+  return <div>NotFoundPage</div>;
+};
+
+export default NotFoundPage;
