+++ src/pages/SignupPage.js
@@ -1,0 +1,5 @@
+const SignupPage = () => {
+  return <div>SignupPage</div>;
+};
+
+export default SignupPage;
